refactor: typescript세팅
