+++ package.json
@@ -13,6 +13,11 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "@types/jest": "^29.2.6",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@types/styled-components": "^5.1.26",
     "animate.css": "^4.1.1",
     "axios": "^0.27.2",
     "dayjs": "^1.11.5",
@@ -39,10 +44,12 @@
     "redux": "^4.2.0",
     "sockjs-client": "^1.6.1",
     "stompjs": "^2.3.3",
-    "styled-components": "^5.3.5",
+    "styled-components": "^5.3.6",
+    "styled-normalize": "^8.0.7",
     "styled-reset": "^4.4.2",
     "sweetalert2": "^11.4.37",
     "swiper": "6.8.4",
+    "typescript": "^4.9.4",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
